Fix for returning only single device
The scan method currently returns a single device, although the scanning will continue for the time specified.
Only a single device will ever be returned this way, which isn't really useful. This patch will change the behaviour of the scan, so that devices are collected and delivered as promised as an array after the scan timeout.